Implement the "instantiate" & "query" entries of "signature-bridge" contract

+++ contracts/signature-bridge/src/state.rs
@@ -7,7 +7,7 @@ use cw_storage_plus::{Item, Map};
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
 pub struct State {
     pub governor: Addr,
-    pub proposal_nonce: u64,
+    pub proposal_nonce: u32,
 }
 
 pub const STATE: Item<State> = Item::new("state");
